+++ mkdocs_markreview/__init__.py
@@ -1,4 +1,0 @@
-from .plugin import MarkReviewPlugin
-
-__all__ = ["__version__", "MarkReviewPlugin"]
-__version__ = "0.1.0"