+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Python package renamed to `mkdocs_markreview`
 ### Removed
 - Deprecated CLI and Docusaurus plugin
+- MkDocs plugin and related examples
 
 ## 0.1.0 - 2024-05-21
 - Initial release with MkDocs and Docusaurus plugins.